requirements.txt
streamlit==1.44.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
boto3==1.34.162
